requirements.txt
agno==1.3.3
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djongo==1.3.7
dnspython==2.7.0
docstring_parser==0.16
exa-py==1.12.1
exceptiongroup==1.2.2
gitdb==4.0.12
GitPython==3.1.44
googlesearch-python==1.3.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
jiter==0.9.0
markdown-it-py==3.0.0
mdurl==0.1.2
openai==1.75.0
packaging==24.2
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
pymongo==3.11.4
pyOpenSSL==25.0.0
pytest==8.3.5
pytest-mock==3.14.0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
service-identity==24.2.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wikipedia==1.4.0
zope.interface==7.2